Figure out the mechanics of the scoring animation in Elm
The key goal is to use CSS animations.

If I translate the JS version directly into Elm the
animation will only play once when the element is first
added to the DOM. I need the animation to play every time
I click the button.

My first thought was to have a list of values and render
that list. That way a new element is added to the DOM
every time I click the button. The problem is that I have
no way of removing the elements that I add.

Custom events to the rescue. All I need to do is listen
for the animationend event on the element and remove it
when it fires.

Since I'm adding and removing elements from the tree Elm
might reuse nodes that should have been removed. If I
click the button in rapid succession then many elements
are added to the tree but not all of them get removed.
To ensure that Elm doesn't mix up nodes I can use
Html.Keyed.

Summary:

  - Use a list of values
  - Attach the animationend event to each rendered element
  - Use Html.Keyed nodes to render each element