Fix a subtle bug concerning how the movement outcome was being handled
On certain movement attempts that don't change the position of the
tiles we were still updating the game in the main model. In these
case Elm wouldn't notice that the game didn't change and Elm would
replay the animation of tiles moving from their last position.

The fix involves never updating the game in the main model if there
was no change to the game. This way Elm doesn't attempt to redraw the
screen. At least that's my take on it.