sed: Apply the replacement to all matches and not just the first
When the build is optimized the CSS is minified and is all on one
line. Since {{ROOT}} needs to be replaced to use the correct path
for the font files we need to replace more that one {{ROOT}} on
the line. Without this change only the first {{ROOT}} gets replaced
which (obviously) causes font loading issues.